refactor(core): moved testType.resolver.ts
